Refactoring for availability form and modal
